@Julian 2021-12-15
Fantastic, thanks!

I knew at least there was some improvement to be made after #817 and #820 changed things, but yeah I think the real issue for #853 will have to do with `_finditem`.